requirements.txt
asgiref==3.5.2
bleach==5.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
django-anymail==8.6
django-environ==0.9.0
django-summernote==0.8.20.0
django-taggit==3.0.0
idna==3.4
Pillow==9.2.0
psycopg2-binary==2.9.3
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.2
urllib3==1.26.12
webencodings==0.5.1
